Create database_quickview.sh
I designed a quick shell script here for previewing any of the databases on our local machines without having to type in all the terminal commands every time. Not included 'Orders' yet as we have not implemented it, but I can easily add it to the script functionality once we do. Cheers S. 

TO RUN: type 'chmod +x database_quickview.sh to turn it into an executable shell script on your system, then you can run it in your terminal using the command './database_quickview.sh'